Support sqlite shell

+++ client/sources/ok_test/sqlite.py
@@ -6,9 +6,16 @@ from client.sources.ok_test import doctest
 from client.utils import format
 from client.utils import timer
 import importlib
+import io
 import os
 import re
 import subprocess
+
+def get_sqlite_shell():
+    sqlite_shell = None
+    try: import sqlite_shell
+    except ImportError: pass
+    return sqlite_shell
 
 class SqliteConsole(interpreter.Console):
     PS1 = 'sqlite> '
@@ -98,13 +105,20 @@ class SqliteConsole(interpreter.Console):
         bool; True if "sqlite3" is executable and the version is at least
         self.VERSION; False otherwise.
         """
-        # Modify PATH in subprocess to check current directory first for sqlite3
-        # executable.
-        try:
-            version = subprocess.check_output(['sqlite3', '--version'],
-                                              env=env).decode()
-        except (subprocess.CalledProcessError, FileNotFoundError):
-            return False
+        args = ['sqlite3', '--version']
+        sqlite_shell = get_sqlite_shell()
+        if sqlite_shell:
+            stdout = io.StringIO()
+            sqlite_shell.main(*args, stdin=io.StringIO(), stdout=stdout, stderr=io.StringIO())
+            version = stdout.getvalue()
+        else:
+            # Modify PATH in subprocess to check current directory first for sqlite3
+            # executable.
+            try:
+                version = subprocess.check_output(args,
+                                                  env=env).decode()
+            except (subprocess.CalledProcessError, FileNotFoundError):
+                return False
         version = version.split(' ')[0].split('.')
         version_info = tuple(int(num) for num in version)
         return version_info >= self.VERSION
@@ -134,18 +148,28 @@ class SqliteConsole(interpreter.Console):
             elif line.startswith(self.PS2):
                 test.append(line[len(self.PS2):])
         test = '\n'.join(test)
-        process = subprocess.Popen(['sqlite3'],
-                                    universal_newlines=True,
-                                    stdin=subprocess.PIPE,
-                                    stdout=subprocess.PIPE,
-                                    stderr=subprocess.PIPE,
-                                    env=env)
-        try:
-            result, error = process.communicate(test, timeout=self.timeout)
-        except subprocess.TimeoutExpired as e:
-            process.kill()
-            print('# Error: evaluation exceeded {} seconds.'.format(self.timeout))
-            raise interpreter.ConsoleException(exceptions.Timeout(self.timeout))
+        args = ['sqlite3']
+        sqlite_shell = get_sqlite_shell()
+        if sqlite_shell:
+            stdin = io.StringIO(test)
+            stdout = io.StringIO()
+            stderr = io.StringIO()
+            sqlite_shell.main(*args, stdin=stdin, stdout=stdout, stderr=stderr)
+            result = stdout.getvalue()
+            error = stderr.getvalue()
+        else:
+            process = subprocess.Popen(args,
+                                        universal_newlines=True,
+                                        stdin=subprocess.PIPE,
+                                        stdout=subprocess.PIPE,
+                                        stderr=subprocess.PIPE,
+                                        env=env)
+            try:
+                result, error = process.communicate(test, timeout=self.timeout)
+            except subprocess.TimeoutExpired as e:
+                process.kill()
+                print('# Error: evaluation exceeded {} seconds.'.format(self.timeout))
+                raise interpreter.ConsoleException(exceptions.Timeout(self.timeout))
         return test, '\n'.join(expected), (error + '\n' + result).strip()
 
 class SqliteSuite(doctest.DoctestSuite):
